fix(#11): Separate collections.py router into collection and file routers
