+++ backend/main.py
@@ -1,6 +1,7 @@
 import os
 import json
 from typing import Dict, Any, List, Optional
+from routers import system, collections_router, files_router
 
 # Load environment variables from .env file
 try:
@@ -94,7 +95,7 @@ app = FastAPI(
 from dependencies import verify_token
 
 # Import routers
-from routers import system, collections
+from routers import system, collections_router, files_router
 
 # Initialize databases on startup
 @app.on_event("startup")
@@ -119,7 +120,8 @@ async def startup_event():
 
 # Include routers
 app.include_router(system.router)
-app.include_router(collections.router)
+app.include_router(collections_router.router)
+app.include_router(files_router.router)
 
 # Configure static files
 static_dir = IngestionService.STATIC_DIR
@@ -162,38 +164,3 @@ async def list_ingestion_plugins(token: str = Depends(verify_token)):
         List of plugin information objects
     """
     return IngestionService.list_plugins()
-
-
-
-@app.get(
-    "/files/{file_id}/content",
-    summary="Get file content",
-    description="Get the content of a file from the collection",
-    tags=["Files"],
-    responses={
-        200: {"description": "File content retrieved successfully"},
-        401: {"description": "Unauthorized - Invalid or missing authentication token"},
-        404: {"description": "File not found"},
-        500: {"description": "Server error"}
-    }
-)
-async def get_file_content(
-    file_id: int,
-    token: str = Depends(verify_token),
-    db: Session = Depends(get_db)
-):
-    """Get the content of a file."""
-    from services.collections import CollectionsService
-    
-    try:
-        return CollectionsService.get_file_content(file_id, db)
-    except HTTPException as e:
-        raise e
-    except Exception as e:
-        import traceback
-        print(f"Error retrieving file content: {str(e)}")
-        print(f"Traceback: {traceback.format_exc()}")
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to retrieve file content: {str(e)}"
-        )
